Third-party update (#84)
* Fix this == 0

* Debugging flag

* Fix this == 0

* unique_ptr instead vector

* Patch tag renaming to (FOnline Patch)

* Memleak fix

* Critter custom name

* Content entries sorting

* Package manager lock aciquiring timeout

* Lighting computation fix

* SDL 2.24.1

* Emscripten 3.1.17

* Using SDL_OpenURL for links

* fmt 9.1.0

* Catch2 3.1.1

* imgui 1.88

* mongo-c-driver 1.23.0

* Setup fixes

* Asio 1.24.0

* Android NDK r25b

* Catch2 reverted to 2.13.10 (v3 not compiled under vs2019)

* Android bson compilation workaround

+++ Source/Tools/ModelBaker.cpp
@@ -561,7 +561,7 @@ static auto FbxGetElement(T* elements, int index, int* vertices) -> T2
         }
     }
 
-    WriteLog("Unknown mapping mode {} or reference mode {}", elements->GetMappingMode(), elements->GetReferenceMode());
+    WriteLog("Unknown mapping mode {} or reference mode {}", static_cast<int>(elements->GetMappingMode()), static_cast<int>(elements->GetReferenceMode()));
     return elements->GetDirectArray().GetAt(0);
 }
